Moved scicrunch into a folder
